improve _sGet and _uGet, convo + chain (almost?) work again
convo + chain haven't been tested too much so that needs to happen but the code is very easy to modify now, and it's not super annoying to work with.
that was a CLEAN recoding
the MessageCreate event now uses a switch case to check through all the server object/user object stuff, makes it looks way nicer n stuff.